requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.18
pytest>=7.4.0
